+++ akeneo-behat/files/pim-initialize.sh
@@ -3,6 +3,7 @@
 echo "Clean the install"
 rm -rf app/archive/*
 rm -rf app/cache/behat
+rm -rf app/cache/test
 rm -rf app/file_storage/*
 rm -rf app/logs/*
 rm -rf web/bundles/*
@@ -15,4 +16,4 @@ app/console -e=behat ca:c
 
 app/console -e=behat oro:requirejs:generate-config
 app/console -e=behat pim:install --force
-app/console -e=behat asset:install --symlink
+app/console -e=behat assets:install --symlink