+++ akeneo-fpm/files/pim-initialize.sh
@@ -13,9 +13,9 @@ rm -rf web/media/*
 echo "Initialize the PIM"
 app/console ca:c -e=prod
 app/console ca:c -e=behat
-app/console ca:c
 
 app/console -e=prod oro:requirejs:generate-config
 app/console -e=prod pim:install --force
-app/console -e=behat pim:install --force
-app/console -e=prod asset:install --symlink
+app/console -e=prod assets:install --symlink
+
+app/console -e=behat pim:installer:db